separated the ISAPI client from the HA integration
