+++ custom_components/hikvision_next/diagnostics.py
@@ -13,7 +13,7 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.device_registry import DeviceEntry
 
 from . import HikvisionConfigEntry
-from .api.const import GET, STREAM_TYPE
+from .isapi.const import GET, STREAM_TYPE
 
 
 def anonymise_mac(orignal: str):
