+++ custom_components/hikvision_next/isapi/const.py
