+++ custom_components/hikvision_next/isapi/models.py
