+++ custom_components/hikvision_next/isapi/utils.py
